+++ corerl/utils/time.py
@@ -6,12 +6,9 @@ def now_iso() -> str:
 
 def split_into_chunks(start: dt.datetime, end: dt.datetime, width: dt.timedelta):
     s = start
-    sec = dt.timedelta(seconds=1)
     while s < end:
         e = s + width
-        # ensure the last bucket does not reach past end
-        e_exclusive = min(end, e - sec)
-        yield(s, e_exclusive)
+        yield(s, min(end, e))
         s = e